+++ pathfinder2e_stats/tests/test_armory.py
@@ -81,14 +81,6 @@ def test_weapons(func):
         armory.pathfinder.melee.macuahuitl,
         armory.pathfinder.ranged.blowgun,
         armory.pathfinder.ranged.dart_umbrella,
-        # scatter weapons
-        armory.pathfinder.ranged.blunderbuss,
-        armory.pathfinder.ranged.dragon_mouth_pistol,
-        armory.pathfinder.ranged.dwarven_scattergun,
-        armory.pathfinder.ranged.explosive_dogslicer,
-        armory.pathfinder.ranged.flingflenser,
-        armory.pathfinder.ranged.spoon_gun,
-        armory.pathfinder.ranged.spraysling,
     ):
         return
 
@@ -128,16 +120,6 @@ def test_macuahuitl():
     )
     assert f(3, 10) == Damage("slashing", 3, 8, 10) + Damage(
         "bleed", 0, 0, 2, persistent=True
-    )
-
-
-def test_spraysling():
-    f = armory.pathfinder.ranged.spraysling
-    assert f() == Damage("bludgeoning", 1, 6) + Damage(
-        "bludgeoning", 0, 0, 1, splash=True
-    )
-    assert f(2, 3) == Damage("bludgeoning", 2, 6, 3) + Damage(
-        "bludgeoning", 0, 0, 2, splash=True
     )